feat: :sparkles: Add category selector and publication adder button
This commit will add the categories of every publication available on website. Installing a new publication will make it displays right on the bottom of it's category

Some publications still have issues on displaying, like monthly Watchtower study edition magazines and Index. Most publications that have a date reference are broken for now