feat: :sparkles: Add "Install JWPUB" button to home view. (not functional)
This commit proposes the change of `sqlite` to `microrm` crate for future work. `microrm` offers a very simple ORM for SQLite based databases, which is the case of `*.jwpub` files. `microrm` is based in `sqlx` while `sqlite` is standalone and API unstable (with a lot of changes e very little version changes), which is harder to maintain. The implementation maybe could create another branch for work, since it will be a complete rework on jwpub lib.